requirements.txt
fastapi
uvicorn
scikit-learn
pandas
transformers
torch
python-dotenv
google-generativeai
joblib
pydantic
